Use name instead of event_name in metric uniqueness in Reporter
Instead of using the `event_name` as a unique_identifier when determining when event is unique, we change it so that it now uses the `name`. The change is needed since in certain case we want to measure the same event_name in multiple ways. This happens when using the keep or old function to discard some events. For example, in the telemetry_metrics hexdoc we have this

```elixir
distribution(
  "http.client.request.duration",
  event_name: [:http_client, :request, :stop],
  drop: &(match?(%{name: :fast_client}, &1))
)

distribution(
  "http.fast.client.request.duration",
  event_name: [:http_client, :request, :stop],
  keep: &(match?(%{name: :fast_client}, &1))
)
```

As we can see, the same event_name is used twice which would be merged together at the moment. Since we need to log them individually, we need to use the `name` instead of `event_name`